refactor: finalize architecture normalization
