+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import { SpeedInsights } from "@vercel/speed-insights/next";
 import { Fragment_Mono, Inter_Tight } from "next/font/google";
 import NextTopLoader from "nextjs-toploader";
 
-import DarkModeProvider from "@/components/DarkModeProvider";
+import DarkModeProvider from "@/app/_providers/DarkModeProvider";
 import "./globals.css";
 
 import type { Metadata } from "next";
