+++ src/features/blog/components/BlogMainSection.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import MotionElement from "@/components/animations/Motion";
+import MotionElement from "@/components/ui/MotionElement";
 
 import type { BlogSummary } from "@/features/blog/types";
