+++ src/features/home/components/PortfolioSection.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 
-import MotionElement from "@/components/animations/Motion";
+import MotionElement from "@/components/ui/MotionElement";
 import SpotifyNowPlaying from "@/features/home/components/SpotifyNowPlaying";
 import GithubActivity from "@/features/home/components/GithubActivity";
 import TechStack from "@/features/home/components/TechStack";
